refactor to directory
